Moved dtos into single file
