--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/budget-dto.interface.ts
@@ -1,8 +1,0 @@
-export interface IBudgetDto {
-   id: number;
-   amount: number;
-   notes: string;
-   monthId: string;
-   name: string;
-   remaining: number;
-}
