--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/budget-list-request-dto.interface.ts
@@ -1,3 +1,0 @@
-export class IBudgetListRequestDto {
-   monthId: string;
-}
