--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/budget-list-response-dto.interface.ts
@@ -1,5 +1,0 @@
-import { IBudgetDto } from '.';
-
-export interface IBudgetListResponseDto {
-   budgets: IBudgetDto[];
-}
