--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/date-range-dto.interface.ts
@@ -1,4 +1,0 @@
-export interface IDateRangeDto {
-   start: Date;
-   end: Date;
-}
