--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/delete-response-dto.interface.ts
@@ -1,3 +1,0 @@
-export interface IDeleteResponseDto {
-   success: boolean;
-}
