--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/id-dto.interface.ts
@@ -1,3 +1,0 @@
-export class IIdDto {
-   id: number;
-}
