--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/index.ts
@@ -1,12 +1,0 @@
-export * from './budget-list-response-dto.interface';
-export * from './budget-list-request-dto.interface';
-export * from './delete-response-dto.interface';
-export * from './login-response-dto.interface';
-export * from './register-request-dto.interface';
-export * from './transaction-list-response-dto.interface';
-export * from './update.response-dto.interface';
-export * from './id-dto.interface';
-export * from './login-request-dto.interface';
-export * from './transaction-dto.interface';
-export * from './budget-dto.interface';
-export * from './date-range-dto.interface';
