--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/login-request-dto.interface.ts
@@ -1,4 +1,0 @@
-export interface ILoginRequestDto {
-   email: string;
-   password: string;
-}
