--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/login-response-dto.interface.ts
@@ -1,6 +1,0 @@
-export interface ILoginResponseDto {
-   success: boolean;
-   error: string;
-   token: string;
-   validTo: string;
-}
