--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/register-request-dto.interface.ts
@@ -1,6 +1,0 @@
-export interface IRegisterRequestDto {
-   email: string;
-   password: string;
-   dateOfBirth: Date;
-   fullName: string;
-}
