--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/transaction-dto.interface.ts
@@ -1,7 +1,0 @@
-export interface ITransactionDto {
-   date: string;
-   description: string;
-   amount: number;
-   budgetIds: number[];
-   id: number;
-}
