--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/transaction-list-response-dto.interface.ts
@@ -1,5 +1,0 @@
-import { ITransactionDto } from '.';
-
-export interface ITransactionListResponseDto {
-   transactions: ITransactionDto[];
-}
