--- MyMoney.Web/ClientApp/src/app/shared/api/dtos/update.response-dto.interface.ts
@@ -1,4 +1,0 @@
-export interface IUpdateResponseDto {
-   success: boolean;
-   error: string;
-}
